webapi/tct-flexiblebox-css3-tests/flexiblebox/CSS3FlexBox_align-self.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        nianx.gan <nianx.gan@intel.com>
        Cui, Jieqiong <jieqiongx.cui@intel.com>

-->

<html>
  <head>
    <title>CSS3 Text Effects Test: CSS3FlexBox_align-self</title>
    <link rel="author" title="Intel" href="http://www.intel.com" />
    <link rel="help" href="https://drafts.csswg.org/css-flexbox-1/#align-items-property" />
    <meta name="flags" content="" />
    <meta name="assert" content="Check if the 'align-self' property exists" />
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
    <script src="support/flexiblebox.js"></script>
    <style>
        #container {
            display: flex;
            align-items: center;
        }

        #no-parent-test {
            display: flex;
        }
    </style>
  </head>
  <body>
    <div id="log"></div>
    <div id="container">
        <div id="has-parent-test"></div>
    </div>
    <div id="no-parent-test"></div>
    <script>
        test(function () {
            var div = document.getElementById("log");
            assert_true(hasStyle("align-self", div.style), "The align-self property exists");
        }, document.title);

        test(function () {
            var div = document.getElementById("has-parent-test");
            var initialValue = getComputedStyle(div)["align-self"];
            assert_equals(initialValue, "center", "The value of align-self property");
        }, "Check that 'align-self' computes the value of 'align-items' on the element's parent "
            + "when the element's 'align-self' set to the default value.");

        test(function () {
            var div = document.getElementById("no-parent-test");
            var initialValue = getComputedStyle(div)["align-self"];
            assert_equals(initialValue, "stretch", "The value of align-self property");
        }, "Check that 'align-self' computes 'stretch' if the element has no parent.");
    </script>
  </body>
</html>
